+++ dcrec/secp256k1/btcec.go
@@ -21,469 +21,6 @@ import (
 	"math/big"
 )
 
-var (
-	// fieldOne is simply the integer 1 in field representation.  It is
-	// used to avoid needing to create it multiple times during the internal
-	// arithmetic.
-	fieldOne = new(fieldVal).SetInt(1)
-)
-
-// addZ1AndZ2EqualsOne adds two Jacobian points that are already known to have
-// z values of 1 and stores the result in (x3, y3, z3).  That is to say
-// (x1, y1, 1) + (x2, y2, 1) = (x3, y3, z3).  It performs faster addition than
-// the generic add routine since less arithmetic is needed due to the ability to
-// avoid the z value multiplications.
-func addZ1AndZ2EqualsOne(x1, y1, z1, x2, y2, x3, y3, z3 *fieldVal) {
-	// To compute the point addition efficiently, this implementation splits
-	// the equation into intermediate elements which are used to minimize
-	// the number of field multiplications using the method shown at:
-	// https://hyperelliptic.org/EFD/g1p/auto-shortw-jacobian-0.html#addition-mmadd-2007-bl
-	//
-	// In particular it performs the calculations using the following:
-	// H = X2-X1, HH = H^2, I = 4*HH, J = H*I, r = 2*(Y2-Y1), V = X1*I
-	// X3 = r^2-J-2*V, Y3 = r*(V-X3)-2*Y1*J, Z3 = 2*H
-	//
-	// This results in a cost of 4 field multiplications, 2 field squarings,
-	// 6 field additions, and 5 integer multiplications.
-
-	// When the x coordinates are the same for two points on the curve, the
-	// y coordinates either must be the same, in which case it is point
-	// doubling, or they are opposite and the result is the point at
-	// infinity per the group law for elliptic curve cryptography.
-	x1.Normalize()
-	y1.Normalize()
-	x2.Normalize()
-	y2.Normalize()
-	if x1.Equals(x2) {
-		if y1.Equals(y2) {
-			// Since x1 == x2 and y1 == y2, point doubling must be
-			// done, otherwise the addition would end up dividing
-			// by zero.
-			doubleJacobian(x1, y1, z1, x3, y3, z3)
-			return
-		}
-
-		// Since x1 == x2 and y1 == -y2, the sum is the point at
-		// infinity per the group law.
-		x3.SetInt(0)
-		y3.SetInt(0)
-		z3.SetInt(0)
-		return
-	}
-
-	// Calculate X3, Y3, and Z3 according to the intermediate elements
-	// breakdown above.
-	var h, i, j, r, v fieldVal
-	var negJ, neg2V, negX3 fieldVal
-	h.Set(x1).Negate(1).Add(x2)                // H = X2-X1 (mag: 3)
-	i.SquareVal(&h).MulInt(4)                  // I = 4*H^2 (mag: 4)
-	j.Mul2(&h, &i)                             // J = H*I (mag: 1)
-	r.Set(y1).Negate(1).Add(y2).MulInt(2)      // r = 2*(Y2-Y1) (mag: 6)
-	v.Mul2(x1, &i)                             // V = X1*I (mag: 1)
-	negJ.Set(&j).Negate(1)                     // negJ = -J (mag: 2)
-	neg2V.Set(&v).MulInt(2).Negate(2)          // neg2V = -(2*V) (mag: 3)
-	x3.Set(&r).Square().Add(&negJ).Add(&neg2V) // X3 = r^2-J-2*V (mag: 6)
-	negX3.Set(x3).Negate(6)                    // negX3 = -X3 (mag: 7)
-	j.Mul(y1).MulInt(2).Negate(2)              // J = -(2*Y1*J) (mag: 3)
-	y3.Set(&v).Add(&negX3).Mul(&r).Add(&j)     // Y3 = r*(V-X3)-2*Y1*J (mag: 4)
-	z3.Set(&h).MulInt(2)                       // Z3 = 2*H (mag: 6)
-
-	// Normalize the resulting field values to a magnitude of 1 as needed.
-	x3.Normalize()
-	y3.Normalize()
-	z3.Normalize()
-}
-
-// addZ1EqualsZ2 adds two Jacobian points that are already known to have the
-// same z value and stores the result in (x3, y3, z3).  That is to say
-// (x1, y1, z1) + (x2, y2, z1) = (x3, y3, z3).  It performs faster addition than
-// the generic add routine since less arithmetic is needed due to the known
-// equivalence.
-func addZ1EqualsZ2(x1, y1, z1, x2, y2, x3, y3, z3 *fieldVal) {
-	// To compute the point addition efficiently, this implementation splits
-	// the equation into intermediate elements which are used to minimize
-	// the number of field multiplications using a slightly modified version
-	// of the method shown at:
-	// https://hyperelliptic.org/EFD/g1p/auto-shortw-jacobian-0.html#addition-mmadd-2007-bl
-	//
-	// In particular it performs the calculations using the following:
-	// A = X2-X1, B = A^2, C=Y2-Y1, D = C^2, E = X1*B, F = X2*B
-	// X3 = D-E-F, Y3 = C*(E-X3)-Y1*(F-E), Z3 = Z1*A
-	//
-	// This results in a cost of 5 field multiplications, 2 field squarings,
-	// 9 field additions, and 0 integer multiplications.
-
-	// When the x coordinates are the same for two points on the curve, the
-	// y coordinates either must be the same, in which case it is point
-	// doubling, or they are opposite and the result is the point at
-	// infinity per the group law for elliptic curve cryptography.
-	x1.Normalize()
-	y1.Normalize()
-	x2.Normalize()
-	y2.Normalize()
-	if x1.Equals(x2) {
-		if y1.Equals(y2) {
-			// Since x1 == x2 and y1 == y2, point doubling must be
-			// done, otherwise the addition would end up dividing
-			// by zero.
-			doubleJacobian(x1, y1, z1, x3, y3, z3)
-			return
-		}
-
-		// Since x1 == x2 and y1 == -y2, the sum is the point at
-		// infinity per the group law.
-		x3.SetInt(0)
-		y3.SetInt(0)
-		z3.SetInt(0)
-		return
-	}
-
-	// Calculate X3, Y3, and Z3 according to the intermediate elements
-	// breakdown above.
-	var a, b, c, d, e, f fieldVal
-	var negX1, negY1, negE, negX3 fieldVal
-	negX1.Set(x1).Negate(1)                // negX1 = -X1 (mag: 2)
-	negY1.Set(y1).Negate(1)                // negY1 = -Y1 (mag: 2)
-	a.Set(&negX1).Add(x2)                  // A = X2-X1 (mag: 3)
-	b.SquareVal(&a)                        // B = A^2 (mag: 1)
-	c.Set(&negY1).Add(y2)                  // C = Y2-Y1 (mag: 3)
-	d.SquareVal(&c)                        // D = C^2 (mag: 1)
-	e.Mul2(x1, &b)                         // E = X1*B (mag: 1)
-	negE.Set(&e).Negate(1)                 // negE = -E (mag: 2)
-	f.Mul2(x2, &b)                         // F = X2*B (mag: 1)
-	x3.Add2(&e, &f).Negate(3).Add(&d)      // X3 = D-E-F (mag: 5)
-	negX3.Set(x3).Negate(5).Normalize()    // negX3 = -X3 (mag: 1)
-	y3.Set(y1).Mul(f.Add(&negE)).Negate(3) // Y3 = -(Y1*(F-E)) (mag: 4)
-	y3.Add(e.Add(&negX3).Mul(&c))          // Y3 = C*(E-X3)+Y3 (mag: 5)
-	z3.Mul2(z1, &a)                        // Z3 = Z1*A (mag: 1)
-
-	// Normalize the resulting field values to a magnitude of 1 as needed.
-	x3.Normalize()
-	y3.Normalize()
-}
-
-// addZ2EqualsOne adds two Jacobian points when the second point is already
-// known to have a z value of 1 (and the z value for the first point is not 1)
-// and stores the result in (x3, y3, z3).  That is to say (x1, y1, z1) +
-// (x2, y2, 1) = (x3, y3, z3).  It performs faster addition than the generic
-// add routine since less arithmetic is needed due to the ability to avoid
-// multiplications by the second point's z value.
-func addZ2EqualsOne(x1, y1, z1, x2, y2, x3, y3, z3 *fieldVal) {
-	// To compute the point addition efficiently, this implementation splits
-	// the equation into intermediate elements which are used to minimize
-	// the number of field multiplications using the method shown at:
-	// https://hyperelliptic.org/EFD/g1p/auto-shortw-jacobian-0.html#addition-madd-2007-bl
-	//
-	// In particular it performs the calculations using the following:
-	// Z1Z1 = Z1^2, U2 = X2*Z1Z1, S2 = Y2*Z1*Z1Z1, H = U2-X1, HH = H^2,
-	// I = 4*HH, J = H*I, r = 2*(S2-Y1), V = X1*I
-	// X3 = r^2-J-2*V, Y3 = r*(V-X3)-2*Y1*J, Z3 = (Z1+H)^2-Z1Z1-HH
-	//
-	// This results in a cost of 7 field multiplications, 4 field squarings,
-	// 9 field additions, and 4 integer multiplications.
-
-	// When the x coordinates are the same for two points on the curve, the
-	// y coordinates either must be the same, in which case it is point
-	// doubling, or they are opposite and the result is the point at
-	// infinity per the group law for elliptic curve cryptography.  Since
-	// any number of Jacobian coordinates can represent the same affine
-	// point, the x and y values need to be converted to like terms.  Due to
-	// the assumption made for this function that the second point has a z
-	// value of 1 (z2=1), the first point is already "converted".
-	var z1z1, u2, s2 fieldVal
-	x1.Normalize()
-	y1.Normalize()
-	z1z1.SquareVal(z1)                        // Z1Z1 = Z1^2 (mag: 1)
-	u2.Set(x2).Mul(&z1z1).Normalize()         // U2 = X2*Z1Z1 (mag: 1)
-	s2.Set(y2).Mul(&z1z1).Mul(z1).Normalize() // S2 = Y2*Z1*Z1Z1 (mag: 1)
-	if x1.Equals(&u2) {
-		if y1.Equals(&s2) {
-			// Since x1 == x2 and y1 == y2, point doubling must be
-			// done, otherwise the addition would end up dividing
-			// by zero.
-			doubleJacobian(x1, y1, z1, x3, y3, z3)
-			return
-		}
-
-		// Since x1 == x2 and y1 == -y2, the sum is the point at
-		// infinity per the group law.
-		x3.SetInt(0)
-		y3.SetInt(0)
-		z3.SetInt(0)
-		return
-	}
-
-	// Calculate X3, Y3, and Z3 according to the intermediate elements
-	// breakdown above.
-	var h, hh, i, j, r, rr, v fieldVal
-	var negX1, negY1, negX3 fieldVal
-	negX1.Set(x1).Negate(1)                // negX1 = -X1 (mag: 2)
-	h.Add2(&u2, &negX1)                    // H = U2-X1 (mag: 3)
-	hh.SquareVal(&h)                       // HH = H^2 (mag: 1)
-	i.Set(&hh).MulInt(4)                   // I = 4 * HH (mag: 4)
-	j.Mul2(&h, &i)                         // J = H*I (mag: 1)
-	negY1.Set(y1).Negate(1)                // negY1 = -Y1 (mag: 2)
-	r.Set(&s2).Add(&negY1).MulInt(2)       // r = 2*(S2-Y1) (mag: 6)
-	rr.SquareVal(&r)                       // rr = r^2 (mag: 1)
-	v.Mul2(x1, &i)                         // V = X1*I (mag: 1)
-	x3.Set(&v).MulInt(2).Add(&j).Negate(3) // X3 = -(J+2*V) (mag: 4)
-	x3.Add(&rr)                            // X3 = r^2+X3 (mag: 5)
-	negX3.Set(x3).Negate(5)                // negX3 = -X3 (mag: 6)
-	y3.Set(y1).Mul(&j).MulInt(2).Negate(2) // Y3 = -(2*Y1*J) (mag: 3)
-	y3.Add(v.Add(&negX3).Mul(&r))          // Y3 = r*(V-X3)+Y3 (mag: 4)
-	z3.Add2(z1, &h).Square()               // Z3 = (Z1+H)^2 (mag: 1)
-	z3.Add(z1z1.Add(&hh).Negate(2))        // Z3 = Z3-(Z1Z1+HH) (mag: 4)
-
-	// Normalize the resulting field values to a magnitude of 1 as needed.
-	x3.Normalize()
-	y3.Normalize()
-	z3.Normalize()
-}
-
-// addGeneric adds two Jacobian points (x1, y1, z1) and (x2, y2, z2) without any
-// assumptions about the z values of the two points and stores the result in
-// (x3, y3, z3).  That is to say (x1, y1, z1) + (x2, y2, z2) = (x3, y3, z3).  It
-// is the slowest of the add routines due to requiring the most arithmetic.
-func addGeneric(x1, y1, z1, x2, y2, z2, x3, y3, z3 *fieldVal) {
-	// To compute the point addition efficiently, this implementation splits
-	// the equation into intermediate elements which are used to minimize
-	// the number of field multiplications using the method shown at:
-	// https://hyperelliptic.org/EFD/g1p/auto-shortw-jacobian-0.html#addition-add-2007-bl
-	//
-	// In particular it performs the calculations using the following:
-	// Z1Z1 = Z1^2, Z2Z2 = Z2^2, U1 = X1*Z2Z2, U2 = X2*Z1Z1, S1 = Y1*Z2*Z2Z2
-	// S2 = Y2*Z1*Z1Z1, H = U2-U1, I = (2*H)^2, J = H*I, r = 2*(S2-S1)
-	// V = U1*I
-	// X3 = r^2-J-2*V, Y3 = r*(V-X3)-2*S1*J, Z3 = ((Z1+Z2)^2-Z1Z1-Z2Z2)*H
-	//
-	// This results in a cost of 11 field multiplications, 5 field squarings,
-	// 9 field additions, and 4 integer multiplications.
-
-	// When the x coordinates are the same for two points on the curve, the
-	// y coordinates either must be the same, in which case it is point
-	// doubling, or they are opposite and the result is the point at
-	// infinity.  Since any number of Jacobian coordinates can represent the
-	// same affine point, the x and y values need to be converted to like
-	// terms.
-	var z1z1, z2z2, u1, u2, s1, s2 fieldVal
-	z1z1.SquareVal(z1)                        // Z1Z1 = Z1^2 (mag: 1)
-	z2z2.SquareVal(z2)                        // Z2Z2 = Z2^2 (mag: 1)
-	u1.Set(x1).Mul(&z2z2).Normalize()         // U1 = X1*Z2Z2 (mag: 1)
-	u2.Set(x2).Mul(&z1z1).Normalize()         // U2 = X2*Z1Z1 (mag: 1)
-	s1.Set(y1).Mul(&z2z2).Mul(z2).Normalize() // S1 = Y1*Z2*Z2Z2 (mag: 1)
-	s2.Set(y2).Mul(&z1z1).Mul(z1).Normalize() // S2 = Y2*Z1*Z1Z1 (mag: 1)
-	if u1.Equals(&u2) {
-		if s1.Equals(&s2) {
-			// Since x1 == x2 and y1 == y2, point doubling must be
-			// done, otherwise the addition would end up dividing
-			// by zero.
-			doubleJacobian(x1, y1, z1, x3, y3, z3)
-			return
-		}
-
-		// Since x1 == x2 and y1 == -y2, the sum is the point at
-		// infinity per the group law.
-		x3.SetInt(0)
-		y3.SetInt(0)
-		z3.SetInt(0)
-		return
-	}
-
-	// Calculate X3, Y3, and Z3 according to the intermediate elements
-	// breakdown above.
-	var h, i, j, r, rr, v fieldVal
-	var negU1, negS1, negX3 fieldVal
-	negU1.Set(&u1).Negate(1)               // negU1 = -U1 (mag: 2)
-	h.Add2(&u2, &negU1)                    // H = U2-U1 (mag: 3)
-	i.Set(&h).MulInt(2).Square()           // I = (2*H)^2 (mag: 2)
-	j.Mul2(&h, &i)                         // J = H*I (mag: 1)
-	negS1.Set(&s1).Negate(1)               // negS1 = -S1 (mag: 2)
-	r.Set(&s2).Add(&negS1).MulInt(2)       // r = 2*(S2-S1) (mag: 6)
-	rr.SquareVal(&r)                       // rr = r^2 (mag: 1)
-	v.Mul2(&u1, &i)                        // V = U1*I (mag: 1)
-	x3.Set(&v).MulInt(2).Add(&j).Negate(3) // X3 = -(J+2*V) (mag: 4)
-	x3.Add(&rr)                            // X3 = r^2+X3 (mag: 5)
-	negX3.Set(x3).Negate(5)                // negX3 = -X3 (mag: 6)
-	y3.Mul2(&s1, &j).MulInt(2).Negate(2)   // Y3 = -(2*S1*J) (mag: 3)
-	y3.Add(v.Add(&negX3).Mul(&r))          // Y3 = r*(V-X3)+Y3 (mag: 4)
-	z3.Add2(z1, z2).Square()               // Z3 = (Z1+Z2)^2 (mag: 1)
-	z3.Add(z1z1.Add(&z2z2).Negate(2))      // Z3 = Z3-(Z1Z1+Z2Z2) (mag: 4)
-	z3.Mul(&h)                             // Z3 = Z3*H (mag: 1)
-
-	// Normalize the resulting field values to a magnitude of 1 as needed.
-	x3.Normalize()
-	y3.Normalize()
-}
-
-// addJacobian adds the passed Jacobian points (x1, y1, z1) and (x2, y2, z2)
-// together and stores the result in (x3, y3, z3).
-func addJacobian(x1, y1, z1, x2, y2, z2, x3, y3, z3 *fieldVal) {
-	// A point at infinity is the identity according to the group law for
-	// elliptic curve cryptography.  Thus, ∞ + P = P and P + ∞ = P.
-	if (x1.IsZero() && y1.IsZero()) || z1.IsZero() {
-		x3.Set(x2)
-		y3.Set(y2)
-		z3.Set(z2)
-		return
-	}
-	if (x2.IsZero() && y2.IsZero()) || z2.IsZero() {
-		x3.Set(x1)
-		y3.Set(y1)
-		z3.Set(z1)
-		return
-	}
-
-	// Faster point addition can be achieved when certain assumptions are
-	// met.  For example, when both points have the same z value, arithmetic
-	// on the z values can be avoided.  This section thus checks for these
-	// conditions and calls an appropriate add function which is accelerated
-	// by using those assumptions.
-	z1.Normalize()
-	z2.Normalize()
-	isZ1One := z1.Equals(fieldOne)
-	isZ2One := z2.Equals(fieldOne)
-	switch {
-	case isZ1One && isZ2One:
-		addZ1AndZ2EqualsOne(x1, y1, z1, x2, y2, x3, y3, z3)
-		return
-	case z1.Equals(z2):
-		addZ1EqualsZ2(x1, y1, z1, x2, y2, x3, y3, z3)
-		return
-	case isZ2One:
-		addZ2EqualsOne(x1, y1, z1, x2, y2, x3, y3, z3)
-		return
-	}
-
-	// None of the above assumptions are true, so fall back to generic
-	// point addition.
-	addGeneric(x1, y1, z1, x2, y2, z2, x3, y3, z3)
-}
-
-// doubleZ1EqualsOne performs point doubling on the passed Jacobian point when
-// the point is already known to have a z value of 1 and stores the result in
-// (x3, y3, z3).  That is to say (x3, y3, z3) = 2*(x1, y1, 1).  It performs
-// faster point doubling than the generic routine since less arithmetic is
-// needed due to the ability to avoid multiplication by the z value.
-func doubleZ1EqualsOne(x1, y1, x3, y3, z3 *fieldVal) {
-	// This function uses the assumptions that z1 is 1, thus the point
-	// doubling formulas reduce to:
-	//
-	// X3 = (3*X1^2)^2 - 8*X1*Y1^2
-	// Y3 = (3*X1^2)*(4*X1*Y1^2 - X3) - 8*Y1^4
-	// Z3 = 2*Y1
-	//
-	// To compute the above efficiently, this implementation splits the
-	// equation into intermediate elements which are used to minimize the
-	// number of field multiplications in favor of field squarings which
-	// are roughly 35% faster than field multiplications with the current
-	// implementation at the time this was written.
-	//
-	// This uses a slightly modified version of the method shown at:
-	// https://hyperelliptic.org/EFD/g1p/auto-shortw-jacobian-0.html#doubling-mdbl-2007-bl
-	//
-	// In particular it performs the calculations using the following:
-	// A = X1^2, B = Y1^2, C = B^2, D = 2*((X1+B)^2-A-C)
-	// E = 3*A, F = E^2, X3 = F-2*D, Y3 = E*(D-X3)-8*C
-	// Z3 = 2*Y1
-	//
-	// This results in a cost of 1 field multiplication, 5 field squarings,
-	// 6 field additions, and 5 integer multiplications.
-	var a, b, c, d, e, f fieldVal
-	z3.Set(y1).MulInt(2)                     // Z3 = 2*Y1 (mag: 2)
-	a.SquareVal(x1)                          // A = X1^2 (mag: 1)
-	b.SquareVal(y1)                          // B = Y1^2 (mag: 1)
-	c.SquareVal(&b)                          // C = B^2 (mag: 1)
-	b.Add(x1).Square()                       // B = (X1+B)^2 (mag: 1)
-	d.Set(&a).Add(&c).Negate(2)              // D = -(A+C) (mag: 3)
-	d.Add(&b).MulInt(2)                      // D = 2*(B+D)(mag: 8)
-	e.Set(&a).MulInt(3)                      // E = 3*A (mag: 3)
-	f.SquareVal(&e)                          // F = E^2 (mag: 1)
-	x3.Set(&d).MulInt(2).Negate(16)          // X3 = -(2*D) (mag: 17)
-	x3.Add(&f)                               // X3 = F+X3 (mag: 18)
-	f.Set(x3).Negate(18).Add(&d).Normalize() // F = D-X3 (mag: 1)
-	y3.Set(&c).MulInt(8).Negate(8)           // Y3 = -(8*C) (mag: 9)
-	y3.Add(f.Mul(&e))                        // Y3 = E*F+Y3 (mag: 10)
-
-	// Normalize the field values back to a magnitude of 1.
-	x3.Normalize()
-	y3.Normalize()
-	z3.Normalize()
-}
-
-// doubleGeneric performs point doubling on the passed Jacobian point without
-// any assumptions about the z value and stores the result in (x3, y3, z3).
-// That is to say (x3, y3, z3) = 2*(x1, y1, z1).  It is the slowest of the point
-// doubling routines due to requiring the most arithmetic.
-func doubleGeneric(x1, y1, z1, x3, y3, z3 *fieldVal) {
-	// Point doubling formula for Jacobian coordinates for the secp256k1
-	// curve:
-	// X3 = (3*X1^2)^2 - 8*X1*Y1^2
-	// Y3 = (3*X1^2)*(4*X1*Y1^2 - X3) - 8*Y1^4
-	// Z3 = 2*Y1*Z1
-	//
-	// To compute the above efficiently, this implementation splits the
-	// equation into intermediate elements which are used to minimize the
-	// number of field multiplications in favor of field squarings which
-	// are roughly 35% faster than field multiplications with the current
-	// implementation at the time this was written.
-	//
-	// This uses a slightly modified version of the method shown at:
-	// https://hyperelliptic.org/EFD/g1p/auto-shortw-jacobian-0.html#doubling-dbl-2009-l
-	//
-	// In particular it performs the calculations using the following:
-	// A = X1^2, B = Y1^2, C = B^2, D = 2*((X1+B)^2-A-C)
-	// E = 3*A, F = E^2, X3 = F-2*D, Y3 = E*(D-X3)-8*C
-	// Z3 = 2*Y1*Z1
-	//
-	// This results in a cost of 1 field multiplication, 5 field squarings,
-	// 6 field additions, and 5 integer multiplications.
-	var a, b, c, d, e, f fieldVal
-	z3.Mul2(y1, z1).MulInt(2)                // Z3 = 2*Y1*Z1 (mag: 2)
-	a.SquareVal(x1)                          // A = X1^2 (mag: 1)
-	b.SquareVal(y1)                          // B = Y1^2 (mag: 1)
-	c.SquareVal(&b)                          // C = B^2 (mag: 1)
-	b.Add(x1).Square()                       // B = (X1+B)^2 (mag: 1)
-	d.Set(&a).Add(&c).Negate(2)              // D = -(A+C) (mag: 3)
-	d.Add(&b).MulInt(2)                      // D = 2*(B+D)(mag: 8)
-	e.Set(&a).MulInt(3)                      // E = 3*A (mag: 3)
-	f.SquareVal(&e)                          // F = E^2 (mag: 1)
-	x3.Set(&d).MulInt(2).Negate(16)          // X3 = -(2*D) (mag: 17)
-	x3.Add(&f)                               // X3 = F+X3 (mag: 18)
-	f.Set(x3).Negate(18).Add(&d).Normalize() // F = D-X3 (mag: 1)
-	y3.Set(&c).MulInt(8).Negate(8)           // Y3 = -(8*C) (mag: 9)
-	y3.Add(f.Mul(&e))                        // Y3 = E*F+Y3 (mag: 10)
-
-	// Normalize the field values back to a magnitude of 1.
-	x3.Normalize()
-	y3.Normalize()
-	z3.Normalize()
-}
-
-// doubleJacobian doubles the passed Jacobian point (x1, y1, z1) and stores the
-// result in (x3, y3, z3).
-func doubleJacobian(x1, y1, z1, x3, y3, z3 *fieldVal) {
-	// Doubling a point at infinity is still infinity.
-	if y1.IsZero() || z1.IsZero() {
-		x3.SetInt(0)
-		y3.SetInt(0)
-		z3.SetInt(0)
-		return
-	}
-
-	// Slightly faster point doubling can be achieved when the z value is 1
-	// by avoiding the multiplication on the z value.  This section calls
-	// a point doubling function which is accelerated by using that
-	// assumption when possible.
-	if z1.Normalize().Equals(fieldOne) {
-		doubleZ1EqualsOne(x1, y1, x3, y3, z3)
-		return
-	}
-
-	// Fall back to generic point doubling which works with arbitrary z
-	// values.
-	doubleGeneric(x1, y1, z1, x3, y3, z3)
-}
-
 // splitK returns a balanced length-two representation of k and their signs.
 // This is algorithm 3.74 from [GECC].
 //
